requirements.txt
boto3==1.39.15
botocore==1.39.15
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
streamlit==1.32.0
PyPDF2==3.0.1
pandas==2.2.0